+++ NEWS.md
@@ -1,5 +1,8 @@
 # Sparklyr 0.7 (UNRELEASED)
 
+- Switched `copy_to` serializer to use Scala implementation, this change can be
+  reverted by setting the `sparklyr.copy.serializer` option to `csv_file`.
+
 - Added support for `spark_web()` for Livy and Databricks connections when
   using Spark 2.X.